move instructions out of edge-split nodes
We keep our pir cfg in edge-split format. What this means is that
every BB can be a merge block, or a split block, but never both.
This avoids pathological graphs like:

       a     b
     /  \  /
    c    d

Where instructions from d cannot be hoisted without being captured by
an unrelated path and instructions from a cannot be pushed down,
without being captured by an unrelated path.

Our cleanup respects this condition. But if a BB is only there to
guarantee edge-split form, we should move all instructions to the
next BB. Many optimizations are BB local, thus it's not a good idea
to have a BB split over two.