fix a bug regarding ldfun
the change in this PR triggered a cornercase, where
scope analysis would change the environment to load from of a ldfun
instruction. but this is not valid, since for ldfun there is the
additional condition that the binding needs to be a closure, otherwise
the environment is skipped and we continue searching in parent envs.